Fixed Warning 'size_t' to 'int'
Changed size return type from int to size_t

+++ lexer/Context.h
@@ -194,7 +194,7 @@ public:
 		return stable.getVariable(t);
 	}
 
-	int printError(string s) {
+	size_t printError(string s) {
 		ofstream outFile;
 		outFile.open(s);
